- improved docs - fixed compiler errors in x64 - conversion from utf8 in CCrashDescReader
git-svn-id: http://crashrpt.googlecode.com/svn@624 821a4644-a21a-0410-aec7-6199d47568cc

+++ branches/1.2/src/crashrptprobe/src/CrashRptProbe.cpp
@@ -144,7 +144,7 @@ int UnzipFile(unzFile hZip, const char* szFileName, const char* szOutFileName)
     if(read_len==0)
       break;
 
-    int written = fwrite(buff, read_len, 1, f);
+    size_t written = fwrite(buff, read_len, 1, f);
     if(written!=1)
       goto cleanup;
   }
